Adding FORRT's description.
I was recommended to this great resource page by Sara Elgebali. I added the traditional/typical FORRT description but if you want smaller, or without hyperlinks in the text, either just cut the last sentence or let me know and I can do it! Thanks a lot for curating this list!